Implemented object pool

+++ Data/PrefabTriangle.cs
@@ -7,10 +7,10 @@ public struct PrefabTriangle
     public Vector2d Position { get; set; }
     public Vector2d Scale { get; set; }
     public double Rotation { get; set; }
-    public int RenderDepth { get; set; }
+    public double RenderDepth { get; set; }
     public int ThemeColor { get; set; }
     
-    public PrefabTriangle(Vector2d position, Vector2d scale, double rotation, int renderDepth, int themeColor)
+    public PrefabTriangle(Vector2d position, Vector2d scale, double rotation, double renderDepth, int themeColor)
     {
         Position = position;
         Scale = scale;
